Zilla-Modules/Dam-Zilla/core/ml/pytorch_integration.py
import torch
import torch.nn as nn
import torch.optim as optim
import numpy as np
import json
from typing import Dict, List, Any
import psycopg2
from psycopg2.extras import RealDictCursor

class ZillaNeuralNetwork(nn.Module):
    def __init__(self, input_size: int, hidden_size: int, output_size: int):
        super(ZillaNeuralNetwork, self).__init__()
        self.layer1 = nn.Linear(input_size, hidden_size)
        self.layer2 = nn.Linear(hidden_size, hidden_size)
        self.layer3 = nn.Linear(hidden_size, output_size)
        self.relu = nn.ReLU()
        self.dropout = nn.Dropout(0.3)
        
    def forward(self, x):
        x = self.relu(self.layer1(x))
        x = self.dropout(x)
        x = self.relu(self.layer2(x))
        x = self.dropout(x)
        x = self.layer3(x)
        return x

class ThreatDetectionModel:
    def __init__(self):
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.model = ZillaNeuralNetwork(50, 128, 5).to(self.device)
        self.optimizer = optim.Adam(self.model.parameters(), lr=0.001)
        self.criterion = nn.CrossEntropyLoss()
        
    def train_model(self, training_data: List[Dict]) -> Dict[str, float]:
        self.model.train()
        total_loss = 0
        
        for batch in training_data:
            inputs = torch.tensor(batch['features'], dtype=torch.float32).to(self.device)
            labels = torch.tensor(batch['labels'], dtype=torch.long).to(self.device)
            
            self.optimizer.zero_grad()
            outputs = self.model(inputs)
            loss = self.criterion(outputs, labels)
            loss.backward()
            self.optimizer.step()
            
            total_loss += loss.item()
            
        return {'average_loss': total_loss / len(training_data)}
    
    def predict_threat(self, features: List[float]) -> Dict[str, float]:
        self.model.eval()
        with torch.no_grad():
            inputs = torch.tensor(features, dtype=torch.float32).to(self.device)
            outputs = self.model(inputs)
            probabilities = torch.softmax(outputs, dim=0)
            
            return {
                'vpn_usage': probabilities[0].item(),
                'burner_number': probabilities[1].item(),
                'puppet_account': probabilities[2].item(),
                'suspicious_activity': probabilities[3].item(),
                'low_risk': probabilities[4].item()
            }

class OpenCVProcessor:
    def __init__(self):
        import cv2
        self.cv2 = cv2
        
    def process_captcha_image(self, image_path: str) -> np.ndarray:
        """Process CAPTCHA images using OpenCV"""
        image = self.cv2.imread(image_path)
        gray = self.cv2.cvtColor(image, self.cv2.COLOR_BGR2GRAY)
        blurred = self.cv2.GaussianBlur(gray, (5, 5), 0)
        edged = self.cv2.Canny(blurred, 50, 150)
        
        return edged
    
    def extract_text_from_image(self, image_path: str) -> str:
        """Extract text from images using OCR with OpenCV preprocessing"""
        processed_image = self.process_captcha_image(image_path)
        # OCR implementation would go here
        return "extracted_text"
