Zilla-Modules/Dam-Zilla/requirements.txt
# === Core Machine Learning & Data Science ===
torch==2.8.0
torchvision==0.16.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.5.0
opencv-python==4.8.1.78
Pillow==10.3.0

# === Web Scraping & Automation ===
requests==2.32.4
beautifulsoup4==4.12.0
selenium==4.11.0
playwright==1.40.0

# === Backend & Task Queue ===
psycopg2-binary==2.9.7
python-dotenv==1.0.0
celery==5.3.0
redis==5.0.0
flower==2.0.0

# === Monitoring ===
prometheus-client==0.17.0

# === Async Networking ===
aiohttp==3.12.14

# === Testing ===
pytest==7.4.0
pytest-asyncio==0.21.0

# === Linting & Formatting (Optional Dev Tools) ===
flake8==6.1.0
black==24.3.0
